+++ src/main/java/com/robertx22/age_of_exile/database/data/spells/spell_classes/bases/MyDamageSource.java
@@ -25,7 +25,6 @@ public class MyDamageSource extends EntityDamageSource {
     static String source(DamageSource s) {
         if (s == null) {
             return DamageEffect.dmgSourceName;
-
         }
         return s.name;
     }