+++ src/main/java/com/robertx22/age_of_exile/mixins/ServerPlayerEntityMixin.java
@@ -1,9 +1,0 @@
-package com.robertx22.age_of_exile.mixins;
-
-import net.minecraft.server.network.ServerPlayerEntity;
-import org.spongepowered.asm.mixin.Mixin;
-
-@Mixin(ServerPlayerEntity.class)
-public class ServerPlayerEntityMixin {
-
-}