+++ src/main/java/com/robertx22/age_of_exile/mmorpg/event_registers/Client.java
@@ -2,9 +2,7 @@ package com.robertx22.age_of_exile.mmorpg.event_registers;
 
 import com.robertx22.age_of_exile.event_hooks.ontick.OnClientTick;
 import com.robertx22.age_of_exile.event_hooks.player.OnKeyPress;
-import com.robertx22.age_of_exile.gui.overlays.bar_overlays.types.HealthReplaceGuiOverlay;
-import com.robertx22.age_of_exile.gui.overlays.bar_overlays.types.RPGPlayerGuiOverlay;
-import com.robertx22.age_of_exile.gui.overlays.bar_overlays.types.VanillaOverlay;
+import com.robertx22.age_of_exile.gui.overlays.bar_overlays.types.RPGGuiOverlay;
 import com.robertx22.age_of_exile.gui.overlays.mob_bar.MobBarScreen;
 import com.robertx22.age_of_exile.gui.overlays.spell_cast_bar.SpellCastBarOverlay;
 import com.robertx22.age_of_exile.gui.overlays.spell_hotbar.SpellHotbarOverlay;
@@ -18,12 +16,10 @@ public class Client {
         ClientTickEvents.END_CLIENT_TICK.register(new OnClientTick());
         ClientTickEvents.END_CLIENT_TICK.register(new OnKeyPress());
 
-        HudRenderCallback.EVENT.register(new VanillaOverlay());
         HudRenderCallback.EVENT.register(new SpellHotbarOverlay());
         HudRenderCallback.EVENT.register(new MobBarScreen());
         HudRenderCallback.EVENT.register(new SpellCastBarOverlay());
-        HudRenderCallback.EVENT.register(new RPGPlayerGuiOverlay());
-        HudRenderCallback.EVENT.register(new HealthReplaceGuiOverlay());
+        HudRenderCallback.EVENT.register(new RPGGuiOverlay());
 
     }
 }