+++ src/main/java/com/robertx22/age_of_exile/uncommon/enumclasses/PlayerGUIs.java
@@ -2,7 +2,5 @@ package com.robertx22.age_of_exile.uncommon.enumclasses;
 
 public enum PlayerGUIs {
     RPG,
-    REPLACE_VANILLA,
-    Vanilla,
     NONE
 }