+++ CHANGELOG.md
@@ -65,5 +65,6 @@ High level players should feel less overpowered now.
 * added spell cast hotkey! Retained right click cast as a config
 * items from craftable utilities mod have been transfered to this mod and are now part of inscribing skill
 * fix items not being enchantable at anvil
+* removed other gui overlay types as this one is configurable enough that it can replace them.